feat(frontend): chat column
Resolve #15830
