+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 -
 
 ### Client
+- Feat: デッキにチャットカラムを追加
 - Fix: ログアウトした際に処理が終了しない問題を修正
 - Fix: 自動バックアップが設定されている環境でログアウト直前に設定をバックアップするように
