+++ packages/frontend/src/deck.ts
@@ -38,6 +38,7 @@ export const columnTypes = [
 	'mentions',
 	'direct',
 	'roleTimeline',
+	'chat',
 ] as const;
 
 export type ColumnType = typeof columnTypes[number];
