+++ packages/frontend/src/pages/chat/home.home.vue
@@ -34,34 +34,7 @@ SPDX-License-Identifier: AGPL-3.0-only
 	<MkFoldableSection>
 		<template #header>{{ i18n.ts._chat.history }}</template>
 
-		<div v-if="history.length > 0" class="_gaps_s">
-			<MkA
-				v-for="item in history"
-				:key="item.id"
-				:class="[$style.message, { [$style.isMe]: item.isMe, [$style.isRead]: item.message.isRead }]"
-				class="_panel"
-				:to="item.message.toRoomId ? `/chat/room/${item.message.toRoomId}` : `/chat/user/${item.other!.id}`"
-			>
-				<MkAvatar v-if="item.message.toRoomId" :class="$style.messageAvatar" :user="item.message.fromUser" indicator :preview="false"/>
-				<MkAvatar v-else-if="item.other" :class="$style.messageAvatar" :user="item.other" indicator :preview="false"/>
-				<div :class="$style.messageBody">
-					<header v-if="item.message.toRoom" :class="$style.messageHeader">
-						<span :class="$style.messageHeaderName"><i class="ti ti-users"></i> {{ item.message.toRoom.name }}</span>
-						<MkTime :time="item.message.createdAt" :class="$style.messageHeaderTime"/>
-					</header>
-					<header v-else :class="$style.messageHeader">
-						<MkUserName :class="$style.messageHeaderName" :user="item.other!"/>
-						<MkAcct :class="$style.messageHeaderUsername" :user="item.other!"/>
-						<MkTime :time="item.message.createdAt" :class="$style.messageHeaderTime"/>
-					</header>
-					<div :class="$style.messageBodyText"><span v-if="item.isMe" :class="$style.youSaid">{{ i18n.ts.you }}:</span>{{ item.message.text }}</div>
-				</div>
-			</MkA>
-		</div>
-		<div v-if="!initializing && history.length == 0" class="_fullinfo">
-			<div>{{ i18n.ts._chat.noHistory }}</div>
-		</div>
-		<MkLoading v-if="initializing"/>
+		<MkChatHistories/>
 	</MkFoldableSection>
 </div>
 </template>
@@ -81,19 +54,11 @@ import { updateCurrentAccountPartial } from '@/accounts.js';
 import MkInput from '@/components/MkInput.vue';
 import MkFoldableSection from '@/components/MkFoldableSection.vue';
 import MkInfo from '@/components/MkInfo.vue';
+import MkChatHistories from '@/components/MkChatHistories.vue';
 
 const $i = ensureSignin();
 
 const router = useRouter();
-
-const initializing = ref(true);
-const fetching = ref(false);
-const history = ref<{
-	id: string;
-	message: Misskey.entities.ChatMessage;
-	other: Misskey.entities.ChatMessage['fromUser'] | Misskey.entities.ChatMessage['toUser'] | null;
-	isMe: boolean;
-}[]>([]);
 
 const searchQuery = ref('');
 const searched = ref(false);
@@ -148,134 +113,14 @@ async function search() {
 	searched.value = true;
 }
 
-async function fetchHistory() {
-	if (fetching.value) return;
-
-	fetching.value = true;
-
-	const [userMessages, roomMessages] = await Promise.all([
-		misskeyApi('chat/history', { room: false }),
-		misskeyApi('chat/history', { room: true }),
-	]);
-
-	history.value = [...userMessages, ...roomMessages]
-		.toSorted((a, b) => new Date(b.createdAt).getTime() - new Date(a.createdAt).getTime())
-		.map(m => ({
-			id: m.id,
-			message: m,
-			other: (!('room' in m) || m.room == null) ? (m.fromUserId === $i.id ? m.toUser : m.fromUser) : null,
-			isMe: m.fromUserId === $i.id,
-		}));
-
-	fetching.value = false;
-	initializing.value = false;
-
-	updateCurrentAccountPartial({ hasUnreadChatMessages: false });
-}
-
-let isActivated = true;
-
-onActivated(() => {
-	isActivated = true;
-});
-
-onDeactivated(() => {
-	isActivated = false;
-});
-
-useInterval(() => {
-	// TODO: DOM的にバックグラウンドになっていないかどうかも考慮する
-	if (!window.document.hidden && isActivated) {
-		fetchHistory();
-	}
-}, 1000 * 10, {
-	immediate: false,
-	afterMounted: true,
-});
-
-onActivated(() => {
-	fetchHistory();
-});
-
 onMounted(() => {
-	fetchHistory();
+	updateCurrentAccountPartial({ hasUnreadChatMessages: false });
 });
 </script>
 
 <style lang="scss" module>
 .start {
 	margin: 0 auto;
-}
-
-.message {
-	position: relative;
-	display: flex;
-	padding: 16px 24px;
-
-	&.isRead,
-	&.isMe {
-		opacity: 0.8;
-	}
-
-	&:not(.isMe):not(.isRead) {
-		&::before {
-			content: '';
-			position: absolute;
-			top: 8px;
-			right: 8px;
-			width: 8px;
-			height: 8px;
-			border-radius: 100%;
-			background-color: var(--MI_THEME-accent);
-		}
-	}
-}
-
-.messageAvatar {
-	width: 50px;
-	height: 50px;
-	margin: 0 16px 0 0;
-}
-
-.messageBody {
-	flex: 1;
-	min-width: 0;
-}
-
-.messageHeader {
-	display: flex;
-	align-items: center;
-	margin-bottom: 2px;
-	white-space: nowrap;
-	overflow: clip;
-}
-
-.messageHeaderName {
-	margin: 0;
-	padding: 0;
-	overflow: hidden;
-	text-overflow: ellipsis;
-	font-size: 1em;
-	font-weight: bold;
-}
-
-.messageHeaderUsername {
-	margin: 0 8px;
-}
-
-.messageHeaderTime {
-	margin-left: auto;
-}
-
-.messageBodyText {
-	overflow: hidden;
-	overflow-wrap: break-word;
-	font-size: 1.1em;
-}
-
-.youSaid {
-	font-weight: bold;
-	margin-right: 0.5em;
 }
 
 .searchResultItem {
